# Conflicts:
#	modules/javafx.controls/src/main/java/javafx/scene/control/skin/MenuButtonSkinBase.java

+++ modules/javafx.base/src/main/java/javafx/util/converter/FloatStringConverter.java
@@ -63,6 +63,6 @@ public class FloatStringConverter extends StringConverter<Float> {
             return "";
         }
 
-        return Float.toString(((Float)value).floatValue());
+        return Float.toString(value.floatValue());
     }
 }
